refactor: Refactor Textarea error
